requirements.txt
appdirs==1.4.4
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
decorator==4.4.2
distlib==0.3.1
dnspython==1.16.0
filelock==3.0.12
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.29
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mstranslate==1.1
numpy==1.19.1
pycparser==2.20
pymongo==3.11.0
python-telegram-bot==12.8
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
tornado==6.0.4
urllib3==1.25.10
virtualenv==20.1.0
